Add files via upload
Here I only created the home page with 5 anchor tags. till now I have just worked in a<section></section> where I added a navigation tag in between a div ,ul tags for the logo and 4 navigation links.
And a another division tag for external text . The best thing I did here was the transition for visit , where I made it in a border with hover effect.